+++ app/src/main/java/com/example/controlefinanceiro/receita/Receita.java
@@ -8,7 +8,7 @@ public class Receita implements Serializable {
 
     private Integer id;
     private String descricao;
-    private float valor;
+    private double valor;
     private String tipo;
     private String categoria;
 
@@ -29,11 +29,11 @@ public class Receita implements Serializable {
         this.descricao = descricao;
     }
 
-    public float getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(float valor) {
+    public void setValor(Double valor) {
         this.valor = valor;
     }
